% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_emp_curve.R
\name{geom_confidence_region}
\alias{geom_confidence_region}
\alias{geom_operating_point}
\title{Add a Confidence Region to an Empirical ROC Curve}
\usage{
geom_confidence_region(
  object,
  cpt,
  ci_level = 0.95,
  colour = "blue",
  linewidth = 0.8
)

geom_operating_point(object, cpt, colour = "red", size = 2.5)
}
\arguments{
\item{object}{an object of class \code{ROC_emp_curve}.}

\item{cpt}{a numeric value specifying the cut-point at which the operating
point and confidence region are computed.}

\item{ci_level}{the confidence level of the confidence region.}

\item{colour}{a character string specifying the colour of the graphical
element.}

\item{linewidth}{a numeric value specifying the line width of the confidence
region boundary.}

\item{size}{a numeric value specifying the point size.}
}
\value{
\code{geom_confidence_region()} returns a \code{ggplot2::geom_contour()} layer.

\code{geom_operating_point()} returns a \code{ggplot2::geom_point()} layer.
}
\description{
Add a Confidence Region to an Empirical ROC Curve

Functions for adding graphical layers to an empirical ROC curve produced by
\code{\link[emplikDTA]{plot.ROC_emp_curve()}}. These functions return standard
 \code{ggplot2} layers and
are intended to be combined with the output of `plot()` using the `+` operator.

\code{geom_confidence_region} adds the confidence region for the pair of
specificity and sensitivity corresponding to a specified cut-point.

\code{geom_operating_point()} adds the empirical operating point
(false positive rate, true positive rate) corresponding to a specified
cut-point.
}
\seealso{
\code{\link[emplikDTA]{ROC_emp_curve()}} and \code{\link[emplikDTA]{plot.ROC_emp_curve()}}
}
