% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_emp_curve.R
\name{ROC_emp_curve}
\alias{ROC_emp_curve}
\alias{ROC_emp_curve.default}
\alias{ROC_emp_curve.formula}
\title{Empirical ROC curve estimation and Empirical Likelihood Inference for ROC curve}
\usage{
ROC_emp_curve(x, ...)

\method{ROC_emp_curve}{default}(x, y, ncp = 150, main, color = "steelblue", plot = TRUE)

\method{ROC_emp_curve}{formula}(formula, data, diag_levels = NULL, subset, na.action, ...)
}
\arguments{
\item{x, y}{numeric vectors (default method) contains values of diagnostic test (biomarker)
corresponding to the first and second class of disease. The ordering is intended 
to be ``increasing'' with respect to the disease severity.}

\item{...}{for formula method: additional arguments to be passed to the
\code{\link[emplikDTA]{ROC_emp_curve}} i.e., \code{ncp}, \code{main}, \code{color}, 
and \code{plot}.}

\item{ncp}{an integer specifying the number of cut-points used to construct
the empirical ROC curve. Larger values produce a smoother approximation at
the expense of additional computation.}

\item{main}{a character string giving the title of the plot.}

\item{color}{a character string specifying the color of the ROC curve.}

\item{plot}{a logical indicating whether the empirical ROC curve should be
plotted. If `FALSE`, the computed ROC curve object is returned without
producing a graphical display.}

\item{formula}{an object of class "\code{\link[stats]{formula}}" (or one that can be coerced 
to that class): a symbolic description of the model to be fitted. 
The details of model specification are given under ‘Details’.}

\item{data}{a data frame containing the variables in the formula.}

\item{diag_levels}{a vector (of strings) containing the ordered name chosen for 
the disease classes. The ordering is intended to be ``increasing'' with respect to the 
disease severity. If \code{diag_levels = NULL} (default), the elements of the vector 
will be automatically determined from data, by considering the order of the means 
of the test values for each disease class (diagnostic group).}

\item{subset}{an optional expression indicating the subset of the rows of data 
that should be used in the fit. This can be a logical vector, or a numeric vector
indicating which observation numbers are to be included, or a character vector of 
the row names to be included. All observations are included by default.}

\item{na.action}{a function which indicates what should happen when the data contain \code{NA}s. 
The default is set by the \code{na.action} setting of \code{\link[base]{options}}, and is 
\code{\link[stats]{na.fail}} if that is unset. The ‘factory-fresh’ default is 
\code{\link[stats]{na.omit}}. Another possible value is \code{NULL}, no action. 
Value \code{\link[stats]{na.exclude}} can be useful.}
}
\value{
\code{ROC_emp_curve} returns an object of class "ROC_emp_curve" which 
is a list containing at least the following components:

\item{call}{the matched call.}
\item{ROCpoint}{the estimated ROC curve: a matrix with two columns named Specificity and
Sensitivity.}
\item{cutpoints}{a vector of cutpoints.}

Generic functions such as \code{plot} is also used to make a plot of ROC surface
 with \code{ggplot2} style.
}
\description{
This function estimates the ROC curve of a continuous diagnostic test (biomarker) 
in two-class setting.
}
\details{
....
}
\references{
To, D. K., Adimari, G., & Chiogna, M. (2024). 
Interval estimation in three-class receiver operating characteristic analysis: 
A fairly general approach based on the empirical likelihood. 
\emph{Statistical Methods in Medical Research}, \bold{33}, 5, 875-893.
}
