% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EL_auc.R
\name{auc}
\alias{auc}
\title{Area Under ROC surface (AUC) estimation and Empirical Likelihood Inference for AUC}
\usage{
auc(x, ...)
}
\arguments{
\item{x, y}{Numeric vectors (default method)}

\item{formula}{A formula of the form y ~ group}

\item{data}{Data frame}

\item{theta0}{Null hypothesis value}

\item{ci_level}{Confidence level}

\item{B}{Bootstrap size}

\item{seed}{Random seed}

\item{plot}{Logical; plot empirical likelihood}
}
\description{
Area Under ROC surface (AUC) estimation and Empirical Likelihood Inference for AUC
}
