% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EL_vus.R
\name{vus}
\alias{vus}
\title{Volume Under Surface (VUS) estimation and Empirical Likelihood Inference for VUS}
\usage{
vus(x, ...)
}
\arguments{
\item{x, y, z}{Numeric vectors (default method)}

\item{formula}{A formula of the form y ~ group}

\item{data}{Data frame}

\item{theta0}{Null hypothesis value}

\item{ci_level}{Confidence level}

\item{B}{Bootstrap size}

\item{seed}{Random seed}

\item{plot}{Logical; plot empirical likelihood}
}
\description{
Volume Under Surface (VUS) estimation and Empirical Likelihood Inference for VUS
}
